move stat def out of localStorage
