+++ src/components/modals/StatsModal.tsx
@@ -21,7 +21,7 @@ import {
   STATISTICS_TITLE,
   VERSE_TEXT_BUTTON,
 } from '../../constants/strings'
-import { GameStats } from '../../lib/localStorage'
+import { GameStats } from '../../constants/types'
 import { shareStatus } from '../../lib/share'
 import { referenceUrl, solutionGameDate, tomorrow } from '../../lib/words'
 import { Histogram } from '../stats/Histogram'
@@ -111,8 +111,8 @@ export const StatsModal = ({
             )}
             {ENABLE_ARCHIVED_GAMES && !isLatestGame && (
               <div className="mt-2 inline-flex">
-                <ClockIcon className="mr-1 mt-2 mt-1 h-5 w-5 stroke-black dark:stroke-white" />
-                <div className="mt-1 ml-1 text-center text-sm sm:text-base">
+                <ClockIcon className="mr-1 mt-1 mt-2 h-5 w-5 stroke-black dark:stroke-white" />
+                <div className="ml-1 mt-1 text-center text-sm sm:text-base">
                   <strong>{ARCHIVE_GAMEDATE_TEXT}:</strong>
                   <br />
                   {format(solutionGameDate, 'd MMMM yyyy', {
@@ -147,13 +147,13 @@ export const StatsModal = ({
       )}
       {ENABLE_MIGRATE_STATS && (
         <div>
-          <hr className="mt-4 -mb-4 border-gray-500" />
+          <hr className="-mb-4 mt-4 border-gray-500" />
           <MigrationIntro handleMigrateStatsButton={handleMigrateStatsButton} />
         </div>
       )}
       {(isGameLost || isGameWon) && (
         <>
-          <hr className="mt-4 -mb-4 border-gray-500" />
+          <hr className="-mb-4 mt-4 border-gray-500" />
           <div className="mt-5 columns-2 items-center items-stretch justify-center text-center dark:text-white sm:mt-6">
             <div className="mt-3 text-xs">
               <p>Read {displayReference} at Bible.com</p>
@@ -171,7 +171,7 @@ export const StatsModal = ({
         </>
       )}
       {/* TODO: Only Show if Signed Out */}
-      <hr className="mt-4 -mb-4 border-gray-500" />
+      <hr className="-mb-4 mt-4 border-gray-500" />
       <div className="mt-5 columns-2 items-center items-stretch justify-center text-center dark:text-white sm:mt-6">
         <div className="mt-3 text-xs">
           <p>Want to save your stats?</p>
