+++ src/constants/types.ts
@@ -1,0 +1,10 @@
+export type GameStats = {
+  winDistribution: number[]
+  gamesFailed: number
+  currentStreak: number
+  bestStreak: number
+  totalGames: number
+  successRate: number
+  score: number
+  avgNumGuesses: number
+}
