+++ src/lib/localStorage.ts
@@ -1,3 +1,5 @@
+import { GameStats } from '@/constants/types'
+
 const gameStateKey = 'gameState'
 const archiveGameStateKey = 'archiveGameState'
 const highContrastKey = 'highContrast'
@@ -22,17 +24,6 @@ export const loadGameStateFromLocalStorage = (isLatestGame: boolean) => {
 }
 
 const gameStatKey = 'gameStats'
-
-export type GameStats = {
-  winDistribution: number[]
-  gamesFailed: number
-  currentStreak: number
-  bestStreak: number
-  totalGames: number
-  successRate: number
-  score: number
-  avgNumGuesses: number
-}
 
 export const saveStatsToLocalStorage = (gameStats: GameStats) => {
   localStorage.setItem(gameStatKey, JSON.stringify(gameStats))
